refactor: SVG 아이콘 컴포넌트를 파일로 분리하고 기존 컴포넌트 제거
