+++ src/components/layout/TheHeader.tsx
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
-import Back from "@/components/UI/Back";
-import Logo from "@/components/UI/Logo";
-import User from "@/components/UI/User";
+import Back from "@/assets/back.svg?react";
+import User from "@/assets/user.svg?react";
+import Logo from "@/components/image/Logo";
 import { useLocation, useNavigate } from "react-router";
 import { ROUTE_PATH } from "@/routes/paths";
 import { useUserInfo } from "@/contexts/UserInfoContext";
@@ -35,13 +35,13 @@ const TheHeader = () => {
   return (
     <Header>
       <button onClick={handleClickBack}>
-        <Back size={"28px"} />
+        <Back width={"28px"} height={"28px"} />
       </button>
       <button onClick={handleClickLogo}>
         <Logo size={"150px"} />
       </button>
       <button onClick={handleClickUser}>
-        <User size={"24px"} />
+        <User width={"24px"} height={"24px"} />
       </button>
     </Header>
   );
